+++ src/submissions/frontend/widgets/app.py
@@ -24,6 +24,7 @@ from .controls_chart import ControlsViewer
 from .kit_creator import KitAdder
 from .submission_type_creator import SubmissionTypeAdder, SubmissionType
 from .sample_search import SearchBox
+from .summary import Summary
 
 logger = logging.getLogger(f'submissions.{__name__}')
 logger.info("Hello, I am a logger")
@@ -69,7 +70,7 @@ class App(QMainWindow):
         fileMenu = menuBar.addMenu("&File")
         # NOTE: Creating menus using a title
         methodsMenu = menuBar.addMenu("&Methods")
-        reportMenu = menuBar.addMenu("&Reports")
+        # reportMenu = menuBar.addMenu("&Reports")
         maintenanceMenu = menuBar.addMenu("&Monthly")
         helpMenu = menuBar.addMenu("&Help")
         helpMenu.addAction(self.helpAction)
@@ -80,7 +81,7 @@ class App(QMainWindow):
         fileMenu.addAction(self.yamlImportAction)
         methodsMenu.addAction(self.searchLog)
         methodsMenu.addAction(self.searchSample)
-        reportMenu.addAction(self.generateReportAction)
+        # reportMenu.addAction(self.generateReportAction)
         maintenanceMenu.addAction(self.joinExtractionAction)
         maintenanceMenu.addAction(self.joinPCRAction)
         
@@ -102,7 +103,7 @@ class App(QMainWindow):
         # logger.debug(f"Creating actions...")
         self.importAction = QAction("&Import Submission", self)
         self.addReagentAction = QAction("Add Reagent", self)
-        self.generateReportAction = QAction("Make Report", self)
+        # self.generateReportAction = QAction("Make Report", self)
         self.addKitAction = QAction("Import Kit", self)
         self.addOrgAction = QAction("Import Org", self)
         self.joinExtractionAction = QAction("Link Extraction Logs")
@@ -122,7 +123,7 @@ class App(QMainWindow):
         # logger.debug(f"Connecting actions...")
         self.importAction.triggered.connect(self.table_widget.formwidget.importSubmission)
         self.addReagentAction.triggered.connect(self.table_widget.formwidget.add_reagent)
-        self.generateReportAction.triggered.connect(self.table_widget.sub_wid.generate_report)
+        # self.generateReportAction.triggered.connect(self.table_widget.sub_wid.generate_report)
         self.joinExtractionAction.triggered.connect(self.table_widget.sub_wid.link_extractions)
         self.joinPCRAction.triggered.connect(self.table_widget.sub_wid.link_pcr)
         self.helpAction.triggered.connect(self.showAbout)
@@ -254,8 +255,8 @@ class AddSubForm(QWidget):
         # NOTE: Add tabs
         self.tabs.addTab(self.tab1,"Submissions")
         self.tabs.addTab(self.tab2,"Controls")
-        self.tabs.addTab(self.tab3, "Add SubmissionType")
-        self.tabs.addTab(self.tab4, "Add Kit")
+        self.tabs.addTab(self.tab3, "Summary Report")
+        # self.tabs.addTab(self.tab4, "Add Kit")
         # NOTE: Create submission adder form
         self.formwidget = SubmissionFormContainer(self)
         self.formlayout = QVBoxLayout(self)
@@ -282,14 +283,15 @@ class AddSubForm(QWidget):
         self.tab2.layout.addWidget(self.controls_viewer)
         self.tab2.setLayout(self.tab2.layout)
         # NOTE: create custom widget to add new tabs
-        ST_adder = SubmissionTypeAdder(self)
+        # ST_adder = SubmissionTypeAdder(self)
+        summary_report = Summary(self)
         self.tab3.layout = QVBoxLayout(self)
-        self.tab3.layout.addWidget(ST_adder)
+        self.tab3.layout.addWidget(summary_report)
         self.tab3.setLayout(self.tab3.layout)
-        kit_adder = KitAdder(self)
-        self.tab4.layout = QVBoxLayout(self)
-        self.tab4.layout.addWidget(kit_adder)
-        self.tab4.setLayout(self.tab4.layout)
+        # kit_adder = KitAdder(self)
+        # self.tab4.layout = QVBoxLayout(self)
+        # self.tab4.layout.addWidget(kit_adder)
+        # self.tab4.setLayout(self.tab4.layout)
         # NOTE: add tabs to main widget
         self.layout.addWidget(self.tabs)
         self.setLayout(self.layout)